==> notebooks/1-process-offers.ipynb <==
# Parsing and Processing Lookup Responses

%load_ext autoreload
%autoreload 2

import os
import glob
import gzip
import json
from datetime import datetime

import multiprocess
import numpy as np
from tqdm import tqdm
import pandas as pd

from config import (
    inc_city_att, 
    inc_city_cl, 
    inc_city_verizon, 
    inc_city_el
)
from parsers import (
    cl_workflow, 
    att_workflow, 
    verizon_workflow, 
    el_workflow, 
    get_incorporated_places, 
    check_redlining, 
    get_holc_grade, 
    get_closest_fiber
)

# inputs
fn_acs = '../data/intermediary/census/aggregated_tables_plus_features.csv.gz'
pattern_att = '../data/intermediary/isp/att/*/*.geojson.gz' # pattern for all data collected from lookup tools
pattern_cl = '../data/intermediary/isp/centurylink/*/*.geojson.gz'
pattern_verizon = '../data/intermediary/isp/verizon/*/*.geojson.gz'
pattern_el = "../data/intermediary/isp/earthlink/*/*.geojson.gz"

# outputs
fn_att = "../data/output/speed_price_att.csv.gz"
fn_cl = '../data/output/speed_price_centurylink.csv.gz'
fn_verizon = '../data/output/speed_price_verizon.csv.gz'
fn_el = '../data/output/speed_price_earthlink.csv.gz'

# params
n_jobs = 20
recalculate = False

# This is from Census data we crunched in the previous notebook.
acs = pd.read_csv(fn_acs, dtype={'geoid': str, 'block_group': str})

# These are the columns we're going to bring to merge with lookup responses.
acs_cols = [
    'geoid', 'race_perc_non_white','income_lmi', 
    'ppl_per_sq_mile', 'n_providers', 'income_dollars_below_median',
    'internet_perc_broadband', 'median_household_income'
]

## Total data collected

def count_addresses(fn):
    """
    How many addresses did we successfully collect in each file?
    """
    import gzip
    import json
    count = 0
    with gzip.open(fn, 'rb') as f:
        for line in f.readlines():
            record = json.loads(line)
            if record['collection_status'] != 0:
                count += 1
    return count 

def count_successful_addresses(pattern, n_jobs=20):
    """
    For all files in `pattern`, sees how many addresses were successfully counted.
    Uses multiprocessing to speed things up.
    """
    files = glob.glob(pattern)
    count = 0
    with multiprocess.get_context("spawn").Pool(n_jobs) as pool:
        for _count in tqdm(pool.imap_unordered(count_addresses, files), 
                           total=len(files)):
            count += _count
    return count

att_count = count_successful_addresses(pattern_att, n_jobs=n_jobs)
verizon_count = count_successful_addresses(pattern_verizon, n_jobs=n_jobs)
cl_count = count_successful_addresses(pattern_cl, n_jobs=n_jobs)
el_count = count_successful_addresses(pattern_el, n_jobs=n_jobs)
all_records = att_count + verizon_count + cl_count + el_count 

print(f"""AT&T: {att_count}
Verizon: {verizon_count}
CenturyLink: {cl_count}
EarthLink: {el_count}
Total: {all_records}""")

## Functions we're going to be using

We `check_redlining` grades by looking if an addresses' coordinates (converted to a Shapely `Point`) are within the `Polygon`s of redlining maps by Mapping Inequality. This actual check is done by `get_holc_grade`.

??get_holc_grade

## ATT

states = []

if not os.path.exists(fn_att) or recalculate:
    # find the data we collected for each block group.
    data_att = []
    files = glob.glob(pattern_att)
    with multiprocess.Pool(n_jobs) as pool:
        # create parallel jobs that parse each block group of data using `att_workflow`.
        for record in tqdm(pool.imap_unordered(att_workflow, files), 
                           total=len(files)):
            data_att.extend(record)
    att = pd.DataFrame(data_att)
    del data_att
    
    # only keep addresses in the incorporated city
    att = att[att.incorporated_place.isin(inc_city_att)]
    att['block_group'] = att['block_group'].apply(lambda x: f"{int(x):012d}")
    
    # check HOLC-grades for each address, and the distance to download speeds at or above 200 Mbps
    att = check_redlining(att)
    # merge census data, and save the file
    att_acs = att.merge(acs[acs_cols], how='left',
                        left_on='block_group', right_on='geoid')
    att_acs = att_acs[[c for c in att_acs.columns if c != 'geoid']]
    att_acs.to_csv(fn_att, index=False, compression='gzip')
else:
    att_acs = pd.read_csv(fn_att)

# start and end collection datetime
[datetime.fromtimestamp(att_acs.collection_datetime.min()), 
 datetime.fromtimestamp(att_acs.collection_datetime.max())]

att_acs['state'].nunique()

len(att_acs)

states.extend(att_acs['state'].unique())

att_acs.redlining_grade.value_counts(normalize=True)

## Centurylink

if not os.path.exists(fn_cl) or recalculate:
    data_cl = []
    files = glob.glob(pattern_cl)
    with multiprocess.Pool(n_jobs) as pool:
        for record in tqdm(pool.imap_unordered(cl_workflow, files), 
                           total=len(files)):
            data_cl.extend(record)
    cl = pd.DataFrame(data_cl)
    del data_cl
    
    cl = cl[cl['incorporated_place'].isin(inc_city_cl)]
    cl = cl[cl.speed_down != 940]
    
    cl = check_redlining(cl)
    cl_acs = cl.merge(acs[acs_cols], how='left', 
                      left_on='block_group', right_on='geoid')
    cl_acs.to_csv(fn_cl, index=False, compression='gzip')
else:
    cl_acs = pd.read_csv(fn_cl)

# start and end collection datetime
[datetime.fromtimestamp(cl_acs.collection_datetime.min()), 
 datetime.fromtimestamp(cl_acs.collection_datetime.max())]

len(cl_acs)

cl_acs['state'].nunique()

states.extend(cl_acs['state'].unique())

cl_acs.redlining_grade.value_counts(normalize=True)

## Verizon

if not os.path.exists(fn_verizon) or recalculate:
    data_verizon = []
    files = glob.glob(pattern_verizon)
    with multiprocess.Pool(n_jobs) as pool:
        for record in tqdm(pool.imap_unordered(verizon_workflow, files), 
                           total=len(files)):
            data_verizon.extend(record)
    verizon = pd.DataFrame(data_verizon)
    del data_verizon
    
    verizon = verizon[verizon.incorporated_place.isin(inc_city_verizon)]
    
    verizon['lon'] = verizon['lon'].astype(float)
    verizon['lat'] = verizon['lat'].astype(float)    
    verizon = check_redlining(verizon)
    
    verizon_acs = verizon.merge(acs[acs_cols], how='left',
                                left_on='block_group', right_on='geoid')
    verizon_acs.to_csv(fn_verizon, index=False, compression='gzip')
else:
    verizon_acs = pd.read_csv(fn_verizon)

# start and end collection datetime
[datetime.fromtimestamp(verizon_acs.collection_datetime.min()), 
 datetime.fromtimestamp(verizon_acs.collection_datetime.max())]

verizon_acs['state'].nunique()

states.extend(verizon_acs['state'].unique())

verizon_acs.redlining_grade.value_counts(normalize=True)

## Earthlink

if not os.path.exists(fn_el) or recalculate:
    data_el = []
    files = glob.glob(pattern_el)
    with multiprocess.Pool(n_jobs) as pool:
        for record in tqdm(pool.imap_unordered(el_workflow, files), 
                           total=len(files)):
            data_el.extend(record)
    el = pd.DataFrame(data_el)
    del data_el
    
    el['block_group'] = el['block_group'].apply(lambda x: f"{int(x):012d}")
    el = check_redlining(el)    
    el_acs = el.merge(acs[acs_cols], how='left', 
                      left_on='block_group', right_on='geoid')
    el_acs = el_acs[el_acs.incorporated_place.isin(inc_city_el)]
    el_acs.to_csv(fn_el, index=False, compression='gzip')
else:
    el_acs = pd.read_csv(fn_el)

# start and end collection datetime
[datetime.fromtimestamp(el_acs.collection_datetime.min()), 
 datetime.fromtimestamp(el_acs.collection_datetime.max())]

datetime.fromtimestamp(el_acs[~el_acs.major_city.isin(['nashville', 'columbus'])].collection_datetime.max())

el_acs['state'].nunique()

# how many states did we collect data from?
states.extend(el_acs['state'].unique())
len(set(states))

# Which ISPs is EarthLink leasing from?
el_acs[el_acs.speed_down != 0].contract_provider.value_counts(dropna=False)

el_acs.incorporated_place.value_counts(normalize=True).head(20)

el_acs.technology.value_counts()

el_acs.redlining_grade.value_counts(normalize=True)
