==> notebooks/4-verizon-spotcheck.ipynb <==
# Verizon Spotcheck
We re-collected Boston addresses and looked at differences in cost and speed.

import glob
from tqdm import tqdm
from multiprocess import Pool

import pandas as pd

from parsers import verizon_workflow

# inputs
fn_verizon = '../data/output/speed_price_verizon.csv.gz'
pattern_spotcheck = '../data/intermediary/isp_spotcheck/verizon/*/*.geojson.gz'

data_verizon = []
files = glob.glob(pattern_spotcheck)
with Pool(20) as pool:
    for record in tqdm(pool.imap_unordered(verizon_workflow, files), 
                       total=len(files)):
        data_verizon.extend(record)
verizon_spot = pd.DataFrame(data_verizon)
del data_verizon

verizon = pd.read_csv(fn_verizon, compression='gzip')

verizon.head(2)

cols = [
    'price_new', 'price_old', 'speed_down_new', 'speed_down_old'
]

verizon_spot = verizon_spot.merge(verizon[['lat', 'lon', 'price', 'speed_down', 'race_perc_non_white', 'income_dollars_below_median']],
                                  how='inner',
                                  on=['lat', 'lon'], suffixes=["_new", "_old"])

verizon_spot[verizon_spot['price_new'] != verizon_spot['price_old']][cols]

verizon_spot.price_new.replace({40: 39.99}, inplace=True)
verizon_spot.price_old.replace({40: 39.99}, inplace=True)

verizon_spot.price_new.describe(), verizon_spot.price_old.describe()

price_changed = verizon_spot[(verizon_spot['price_new'] != verizon_spot['price_old']) &
             (~verizon_spot.price_new.isnull()) &
             (~verizon_spot.price_old.isnull())]

price_changed[cols]

len(price_changed[
    price_changed['price_new'] > price_changed['price_old']
]) / len(price_changed)

len(price_changed[
    price_changed['price_new'] < price_changed['price_old']
]) / len(price_changed)

# note none of the speeds change.
speed_changed = verizon_spot[
    (verizon_spot['speed_down_new'] != verizon_spot['speed_down_old']) &
    (~verizon_spot['price_new'].isnull()) &
    (~verizon_spot['price_old'].isnull())
]

len(verizon_spot[verizon_spot.speed_down_old <= 25])

len(speed_changed)

speed_changed[speed_changed['speed_down_new'] == 300][cols]

speed_changed[speed_changed['speed_down_old'] == 300][cols]

speed_changed['speed_down_old'].value_counts()

all but one instance of a household being offered different speeds was for Verizon's "High Speed Internet"

verizon_spot[verizon_spot['price_new'] > 39.99].income_dollars_below_median.describe()

verizon_spot[verizon_spot['price_old'] > 39.99].income_dollars_below_median.describe()

verizon_spot[verizon_spot['price_new'] > 39.99].race_perc_non_white.describe()

verizon_spot[verizon_spot['price_old'] > 39.99].race_perc_non_white.describe()

# median income for everything
verizon_spot.income_dollars_below_median.describe()

verizon_spot.race_perc_non_white.describe()

The median income was lower in higher cost plans from the first sample, but not the second sample.

The racial demographics were less white for more expensive plans

verizon_spot[verizon_spot.price_old != '39.99'].price_new.value_counts(normalize=True)

verizon_spot[verizon_spot.price_new != '39.99'].price_old.value_counts(normalize=True)

95 percent of offers that were over 39.99 were switched back to 39.99 when we checked again. Similarly, 77 percent of offers that are now over 39.99 were originally 39.99.
