==> notebooks/5-find-closest-fiber-address.ipynb <==
# Finding the closes household with Fiber

%load_ext autoreload
%autoreload 2

import pandas as pd
from parsers import get_closest_fiber
from aggregators import filter_df

# inputs
fn_att = '../data/output/speed_price_att.csv.gz'
fn_centurylink = '../data/output/speed_price_centurylink.csv.gz'
fn_verizon = '../data/output/speed_price_verizon.csv.gz'
fn_earthlink = '../data/output/speed_price_earthlink.csv.gz'
inputs = {
    "AT&T" : fn_att,
    "CenturyLink": fn_centurylink,
    "Verizon": fn_verizon,
    "EarthLink" : fn_earthlink
}

??get_closest_fiber

view_cols = [
    'address_full',
    'address_full_closest_fiber',
    'closest_fiber_miles',
    'speed_down',
    'median_household_income',
    'median_household_income_closest_fiber',
    'race_perc_non_white',
    'race_perc_non_white_closest_fiber'
    
]

for isp, fn in inputs.items():
    df = filter_df(fn, isp)
    break

for city, df__ in df.groupby('major_city'):
    df__ = get_closest_fiber(df__.reset_index(drop=True))
    print(df__[df__.speed_down < 200][view_cols].sort_values('closest_fiber_miles').head(3))
    break

# get the three closest address pairs where an address in a low income area is closest to a blazing fast offer to an upper income area.
df__[(df__.income_level == 'Low') & 
     (df__.income_level_closest_fiber == 'Upper Income')].sort_values('closest_fiber_miles')[view_cols].head(3)
